Move error reporting back to semantic. Add missing import
